Fix Diagnostics Issues
1) Fix diagnostics issues - at the server level, at the cluster level
2) Remove redundant code from system_server
3) Fix RPC APIs
